audit(demo): CDemo::Render was a signedness bug, not a wall - 98.91%->100.00% EXACT
The m_520 idle-timer compare (g_frameDelta >= m_520) emitted signed jl where
retail uses unsigned jb; casting the compare to u32 makes cl emit the unsigned
branch. Asm now byte-identical (55 insns); the prior @early-stop mislabeled it a
regalloc coin-flip + IAT artifact.

+++ src/Gruntz/Demo.cpp
@@ -83,11 +83,11 @@ i32 CWorldState::BuildWorldLevelKey(i32 unused) {
 // idle timer m_520 down by the frame delta, and posts WM_COMMAND 0x8027 when it
 // expires. Returns 1.
 // @early-stop
-// 99.55%: every opcode/offset/branch is byte-identical. The residual is (1) the
-// PostMessageA IAT-absolute scoring artifact (target bakes the bare 0x6c44c8, no
-// symbol) and (2) a register-coloring coin-flip (pm <-> ebx/edi vs the 0x100 mask) -
-// the documented regalloc back-edge wall (docs/patterns/zero-register-pinning.md).
-// Not source-steerable; deferred to the final sweep.
+// asm now byte-IDENTICAL (55 insns): the prior residual included a real signedness
+// bug - the m_520 timer compare emitted signed jl where retail uses unsigned jb;
+// fixed by casting the g_frameDelta>=m_520 test to u32. The only remaining residual
+// is the PostMessageA IAT-absolute scoring artifact (target bakes the bare 0x6c44c8,
+// no symbol) - reloc-masked truth, not a code diff. docs/matching-patterns.md.
 RVA(0x0003c220, 0xa4)
 i32 CDemo::Render() {
     CPlay::Render();
@@ -99,7 +99,7 @@ i32 CDemo::Render() {
             break;
         }
     }
-    if (g_frameDelta >= m_520) {
+    if (static_cast<u32>(g_frameDelta) >= static_cast<u32>(m_520)) {
         m_520 = 0;
     } else {
         m_520 -= g_frameDelta;
